requirements.txt
# === SCRAPING ===
apify-client==1.7.1
requests==2.31.0
beautifulsoup4==4.12.2

# === DATA PROCESSING ===
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2

# === AUTOMATION ===
schedule==1.2.0
python-dotenv==1.0.0

# === UTILITIES ===
aiofiles==23.2.1
